Partially convert lastminute script from WF-Moscow
Explain defaults and directly overwrite in vars.
The idea is to use ansible roles which can be added to multiple
playbooks, for example the "backup contestant machines (in case of
failure) the main contestant machines.

Also use default linter
This should prevent us making major mistakes in forgetting to set the
correct values etc.