Fix script-server to work with new inventory
Our new inventory file doesn't include ansible_host, since we're just
letting dns(i.e. the /etc/hosts file) resolve our system names. Because
of this we need to lookup the ips for these hostnames in order to add
them to script server as allowed users.

Also change to just plain ansible-playbook, because we don't have
fast-ansible-playbook anymore; it was an experiment in Astana, but
lost/maybe not really necessary.