Update packages role to be a passthrough proxy/cache
Caches files, but fetches them from upstream if they're missing. So if
we have a full mirror, it only serves up locally, otherwise fetches from
the upstream server.

This makes it great when running vms locally for development, because
you don't need the full mirror in order to point clients at packages,
and you get the benefit of files being cached.

Also restart nginx instead of reload, because reload doesn't seem to be
sufficient to get the config fully loaded properly.