Make sure extra files are set up for cds + fix presentations
Out of the box, the wlp server includes some extra files that we weren't
ensuring were present for servers named other than cds. This fixes that
oversight, just in case there were any behavior differences with these
files present.

Also make sure that the presentations directory is properly a symlink,
otherwise the task fails because it can't replace a directory with a
symlink.