A whole pile of changes for Baku (#149)
* Bootstrap the git repos that the gitserver should have

We run it as a local mirror/working copy of upstream repos, so we should
start off with a copy of the remotes so it's functional out of the box.

* Fix script-server to work with new inventory

Our new inventory file doesn't include ansible_host, since we're just
letting dns(i.e. the /etc/hosts file) resolve our system names. Because
of this we need to lookup the ips for these hostnames in order to add
them to script server as allowed users.

Also change to just plain ansible-playbook, because we don't have
fast-ansible-playbook anymore; it was an experiment in Astana, but
lost/maybe not really necessary.

* Remove things that have been fixed upstream from lastminute

And fix one command that was using the wrong shell so it wasn't working
properly.

* Update packages role to be a passthrough proxy/cache

Caches files, but fetches them from upstream if they're missing. So if
we have a full mirror, it only serves up locally, otherwise fetches from
the upstream server.

This makes it great when running vms locally for development, because
you don't need the full mirror in order to point clients at packages,
and you get the benefit of files being cached.

Also restart nginx instead of reload, because reload doesn't seem to be
sufficient to get the config fully loaded properly.

* Allow extra iptables rules via ansible variables

We already do this for backup, and it helps simplify some local
development work.

* Make sure extra files are set up for cds + fix presentations

Out of the box, the wlp server includes some extra files that we weren't
ensuring were present for servers named other than cds. This fixes that
oversight, just in case there were any behavior differences with these
files present.

Also make sure that the presentations directory is properly a symlink,
otherwise the task fails because it can't replace a directory with a
symlink.

* Add a `github_release` helper module + use it for the cds

This module makes it a bit easier to fetch assets from a github release,
including caching. It lets us replace a whole bunch of ugly/conditional
ansible with a single task.

* Remove strict hostkey checking from cds role

We don't need it, since we're going to add ssh host certificates and
trust everything automatically that way.

* Disable nginx to free up ports 80/443 for the cds

Otherwise it'll fail to start properly

* Add an ssh_ca role

This takes care of generating an ssh host and user ca, host
certificates, and a default icpc@hostname key+certificate. Hosts also
get set up to trust the user ssh ca, so they can put things like this in
~/.ssh/authorized_principals and then the icpc account on the
corresponding machine will be able to ssh in:
```
icpc@cds
icpc@backup
```

I had to slightly re-organize lastminute to make sure things happen in
the right order (the ssh ca needs to be created before we try to use it)

* Give the cds an ssh key certificate

We'll use this to let it log into the gitserver, and potentially others.

* Add a bunch of trusted users that can access the gitserver

Using our newly minted ssh user certificates, we can grant access from a
bunch of hosts to the gitserver.

* Set some host_vars defaults for the cds

Without these, it'll blow up and error.

* Update hosts file on backup

Made this more or less match the document here:
https://docs.google.com/document/d/1DUgmZWQlmVEw7QzvDpZGr3oVmn_KltYdFZEBPv7Smb0/edit?tab=t.0

The only changes are commenting out things that weren't hostnames, or
renaming a few things to be valid hostnames.

* Resolve linter issues

* Handle root_password: none the same as if it was undefined

Otherwise ansible errors out

* Make sure update-ca-certificates gets run

The image was built with the certs installed in the right place, but we
never called update-ca-certificates so they're not actually usable out
of the box. We need to make sure this happens, or they won't be able to
install packages/trust the cds/etc.